requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.3
